Move get_exporter_config helper to utils
